refactor: relocate tree item components
